Remove exceptions and make Instruction constexpr
This change replaces the throw in the default cases for Register and
Operator switches with an unreachable annotation.  This will cause a
compiler warning if a new enumerator is added to either enumeration,
rather than a runtime error.

This change also removes the exception thrown when an Instruction fails
to decode.  Instead of throwing an exception, Instruction::decode now
returns an optional value.  This makes it possible for Instruction
decoding to be constexpr.