Add machine::memory_size constant
This change adds the machine::memory_size constant, which is the size
of Memory in bytes.  The size member of std::array is non-static, so
the size of an array type can only be determined by constructing a value
of that array type and querying its size.  This process is refactored
into the memory_size constant because it can be awkward to read.