Add machine::create_program function
This change adds an overload of machine::load_into_memory for loading
into memory from an iterator begin/end pair.  This overload is used in
the new equivalent overload of machine::load_program.

These functions are used in the new create_program function, which is
can statically construct a memory array from a simple array with
guaranteed success.  This is achieved by checking the bounds of the
simple array against the size of the memory array.  This will be useful
for quickly creating memory arrays containing simple programs in test
cases for the virtual machine.