Fix Windows build and test failures
This change addresses several compile errors and warnings on Windows and
makes several adjustments so that every test passes.  The main change is
that the memory and display arrays are now allocated on the heap, as the
default stack size of 16KiB on Windows was being exhausted while running
the machine tests.