Do not modify program counter during fetch
Fetch no longer updates the program counter.  Instead, each instruction
is responsible for modifying the program counter directly.  This also
allows the typically-unused pc instruction for each execute method to be
removed.  As a side effect, the program counter after a fault no longer
points to the instruction after the faulting instruction.