Remove LIBNPLN_DETAIL_UNREACHABLE
The unreachable annotation was being incorrectly used to write total
noexcept functions where the body was an assumed exhaustive switch over
an enum.  This would result in undefined behavior when an invalid
enumerator was switched over.  Such enumerators can be produced by
typecasting an integral value not represented by an enumerator into the
enum type.

This change corrects this by throwing after the switch body instead of
using the unreachable annotation to suppress the compiler warning.  This
requires removing the noexcept specifier from these functions.