Add range checking for nibble in drw_v_v_n
This range check ensures that the nibble operand is not outside of the
acceptable range.

This nibble is obtained by decoding an instruction.  The decoder should
never produce a value that cannot fit in a nibble.  execute_drw_v_v_n is
private and can only be called indirectly by inserting a drw_v_v_n
instruction into memory.  Clients of Machine thus cannot provide an
invalid nibble to this method, but range checking should be done
regardless.