Disable readability-function-cognitive-complexity check
This check does not tolerate functions generated by Catch2.  It has an
option, IgnoreMacros, that partially alleviates this but the resulting
functions generated by those macros still fail the check.  Ideally, it
would be possible to specify an expression for function names to be
ignored by the check.  Alternatively, if the set of checks could be
modified for files matching an expression (such as .test.cpp files),
this check could be disabled on a per-file basis without requiring
modification of the files.