Integrate Conan package manager
Conan installs the necessary dependencies for building the project.  It
uses a package list file that gives the developer precise control over
the package versions being used.  It integrates with CMake to use the
project-local repository of packages to resolve find_package searches.

This change updates the following dependency versions:
  - CMake 3.21
  - Catch 2.13.7
  - CLI11 2.1.1
  - GLFW 3.3.4
  - fmtlib 8.0.1

This change removes the following vendor dependencies:
  - Catch
  - Glad

GSL is kept as a vendor dependency.  The latest release of GSL is 3.1.0,
tagged in June 2020.  This project requires several changes made since
then, such as non_null<unique_ptr> and narrow.